Fix Lenovo lookup and catch errors
It's currently necessary to make requests to a URL of the form:

```
Laptops-and-netbooks/ThinkPad-T-Series-laptops/ThinkPad-T430si/2353/9KU
```

rather than

```
Laptops-and-netbooks/ThinkPad-T-Series-laptops/ThinkPad-T430si/2353
```

The former is provided under the `Id` key in the `url` request.

Also, it's possible that the result will be `null`, rather than valid
JSON. (This is currently the case when using the `ParentId` value).
To deal with this (and other potential related errors), a
try/catch block around the `json.loads()` block is required.